pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gsfs"
version = "0.1.0"
description = "GSfS - RSS Feed Collector and Processor"
authors = [
    {name = "dmc", email = "davidmjchoi@outlook.com"},
]

dependencies = [
    "feedparser>=6.0.10",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "markdown>=3.5.1",
    "beautifulsoup4>=4.12.2",
    "playwright==1.56.0",
    "openai>=2.8.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black",
    "flake8",
]

[tool.setuptools.packages.find]
where = ["."]  # 在当前目录查找包
include = ["src*", "src/scraper*"]  # 包含 src 目录下的所有包